moved js code related to jetpack integration into the custom integration component
